Add sort + reorder mode to the mobile subscriptions sheet
The subscriptions tree (sort menu, drag-to-reorder) only existed on the
desktop sidebar; mobile's SubscriptionsSheet still had the old read-only
accordion list. Extracted the tree's data-building (useSubscriptionsTree)
and rendering (SubscriptionsTree) into shared modules so both surfaces
use one implementation instead of two divergent ones — the desktop
sidebar computes the tree once and passes the same topNodes/groupNodes/
dragAndDropHooks down to the sheet, which mounts its own <Tree> from
that identical data (react-aria's dragAndDropHooks is a stateless
hook-factory bag, safe to share across two <Tree> instances).

The mobile sheet now has the same overflow menu (Sort submenu, Reorder
subscriptions toggle with the settings-gear/checkmark swap, collapse/
expand all) and the same drag-and-drop tree (preview, drop-line
indicator, drop-target highlight) as desktop. reorderMode is one shared,
unpersisted toggle in AppShell rather than a per-surface copy, since
only one of the two tree mounts is ever visible at a given viewport
width. Removed the old SheetPubRow/SheetUserRow/SheetListGroup Disclosure
rendering and the dead, never-wired onReorder prop.
